# conditionally require `FreeType`, `FileIO`, `Unitful` and `ImageInTerminal` through weak deps or `Requires`