--- a/ext/FreeTypeExt.jl
+++ b/ext/FreeTypeExt.jl
@@ -1,13 +1,12 @@
 # adapted from github.com/JuliaGraphics/FreeTypeAbstraction.jl/blob/master/src/rendering.jl
 # credits to the `FreeTypeAbstraction` authors
 
-module FreeTypeRendering
+module FreeTypeExt
 
-using ..StaticArrays
-using ..ColorTypes
-using FreeType
-
-export get_font_face, render_string!
+import UnicodePlots
+isdefined(Base, :get_extension) ? (using FreeType) : (using ..FreeType)    
+using StaticArrays
+using ColorTypes
 
 const REGULAR_STYLES = "regular", "normal", "medium", "standard", "roman", "book"
 const FT_LIB = FT_Library[C_NULL]
@@ -85,7 +84,7 @@ fallback_fonts() =
 
 const FT_FONTS = Dict{String,FTFont}()
 
-function get_font_face(font = nothing, fallback = fallback_fonts())
+function UnicodePlots.get_font_face(font = nothing, fallback = fallback_fonts())
     face = nothing
     for name ∈ filter(!isnothing, (font, "JuliaMono", fallback...))
         if (face = get(FT_FONTS, name, nothing)) ≡ nothing
@@ -287,7 +286,7 @@ Uses the conventions of freetype.org/freetype2/docs/glyphs/glyphs-3.html
 * `gstr`: background string or array of chars (for background sizing)
 * `incx`: extra x spacing
 """
-function render_string!(
+function UnicodePlots.render_string!(
     img::AbstractMatrix{T},
     fstr::Union{AbstractVector{Char},String},
     face::FTFont,