--- a/src/UnicodePlots.jl
+++ b/src/UnicodePlots.jl
@@ -18,7 +18,6 @@ import ColorSchemes
 import Requires
 import NaNMath
 import Contour
-import FileIO
 
 # composite types
 export Plot,
@@ -127,9 +126,11 @@ function __init__()
         Requires.@require ImageInTerminal = "d8c32880-2388-543b-8c61-d9f865259254" include(
             "../ext/ImageInTerminalExt.jl",
         )
-        Requires.@require FreeType = "b38be410-82b0-50bf-ab77-7b57e271db43" include(
-            "../ext/FreeTypeExt.jl",
-        )
+        Requires.@require FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549" begin
+            Requires.@require FreeType = "b38be410-82b0-50bf-ab77-7b57e271db43" begin
+                include("../ext/FreeTypeExt.jl")
+            end
+        end
     end
     nothing
 end