--- a/src/interface/scatterplot.jl
+++ b/src/interface/scatterplot.jl
@@ -132,32 +132,3 @@ function scatterplot!(plot::Plot{<:Canvas}, x::AbstractVector, y::AbstractMatrix
     end
     plot
 end
-
-# ---------------------------------------------------------------------------- #
-# Unitful
-function scatterplot(
-    x::AbstractVector{<:RealOrRealQuantity},
-    y::AbstractVector{<:Quantity};
-    unicode_exponent::Bool = KEYWORDS.unicode_exponent,
-    xlabel = KEYWORDS.xlabel,
-    ylabel = KEYWORDS.ylabel,
-    kw...,
-)
-    x, ux = number_unit(x, unicode_exponent)
-    y, uy = number_unit(y, unicode_exponent)
-    scatterplot(
-        ustrip.(x),
-        ustrip.(y);
-        xlabel = unit_label(xlabel, ux),
-        ylabel = unit_label(ylabel, uy),
-        unicode_exponent,
-        kw...,
-    )
-end
-
-scatterplot!(
-    plot::Plot{<:Canvas},
-    x::AbstractVector{<:RealOrRealQuantity},
-    y::AbstractVector{<:Quantity};
-    kw...,
-) = scatterplot!(plot, ustrip.(x), ustrip.(y); kw...)