--- a/test/tst_common.jl
+++ b/test/tst_common.jl
@@ -285,14 +285,20 @@ end
     @test UnicodePlots.keywords() isa String
 end
 
+const UE = if isdefined(Base, :get_extension)
+    Base.get_extension(UnicodePlots, :UnitfulExt)
+else
+    UnicodePlots.UnitfulExt
+end
+
 @testset "units" begin
     x = [1.0, 2.0, 3.0]
-    @test UnicodePlots.number_unit(x) == (x, nothing)
-    @test UnicodePlots.number_unit(1) == (1, nothing)
-    @test UnicodePlots.number_unit(x * u"°C") == (x, "°C")
-    @test UnicodePlots.number_unit(1u"°C") == (1, "°C")
+    @test UE.number_unit(x) == (x, nothing)
+    @test UE.number_unit(1) == (1, nothing)
+    @test UE.number_unit(x * u"°C") == (x, "°C")
+    @test UE.number_unit(1u"°C") == (1, "°C")
 
-    @test UnicodePlots.unit_label("  fancy label  ", "Hz") == "  fancy label (Hz)"
-    @test UnicodePlots.unit_label("  ", "dB") == "dB"
-    @test UnicodePlots.unit_label("  no units  ", nothing) == "  no units"
+    @test UE.unit_label("  fancy label  ", "Hz") == "  fancy label (Hz)"
+    @test UE.unit_label("  ", "dB") == "dB"
+    @test UE.unit_label("  no units  ", nothing) == "  no units"
 end