--- a/test/tst_quality.jl
+++ b/test/tst_quality.jl
@@ -1,5 +1,5 @@
 @testset "Aqua" begin
     # JuliaTesting/Aqua.jl/issues/77
-    Aqua.test_all(UnicodePlots; ambiguities = false)
+    Aqua.test_all(UnicodePlots; ambiguities = false, project_toml_formatting = false)
     Aqua.test_ambiguities(UnicodePlots)
 end