… .github/workflows/format-check.yml
… CHANGELOG.md
… Project.toml
… README.md
… test/tst_freetype.jl
… test/tst_io.jl