Fix https://github.com/JuliaPlots/UnicodePlots.jl/issues/331.
Saving as `.png` is mostly for `Plots`, and not the primary target of `UnicodePlots` so we now make `FreeType` and `FileIO` optional.
It's also likely `Unitful` is already loaded when you are plotting quantities, so it's unlikely that making it a weak dependency will break anyone's workflow (glue code) even if this is technically breaking ?

~~I guess this will require a major bump ... Thoughts ?~~

As mentioned on slack, making glue package code optional seems to be non breaking in this case so a minor bump is probably sufficient.

I think I'll wait the `1.9.0-beta4` release to merge this PR for the backport of https://github.com/JuliaLang/julia/pull/48352.

<details>
<summary>latency improvements: ~2s, about 32% gain </summary>

`julia 1.9.0-beta3`

**pr**
```julia
julia> @time @time_imports using UnicodePlots
     30.5 ms  Preferences
      0.3 ms  SnoopPrecompile
      0.9 ms  Statistics
      3.3 ms  StaticArraysCore
    885.5 ms  StaticArrays
     61.0 ms  FixedPointNumbers
    140.2 ms  ColorTypes 21.49% compilation time (100% recompilation)
     38.9 ms  Crayons
      1.2 ms  DataAPI
      0.2 ms  Compat
     10.2 ms  OrderedCollections
    106.7 ms  DataStructures
      0.5 ms  SortingAlgorithms
     12.6 ms  Missings
      1.4 ms  DocStringExtensions
      7.1 ms  IrrationalConstants
      0.4 ms  LogExpFunctions
      0.3 ms  StatsAPI
     33.6 ms  StatsBase
     81.4 ms  MarchingCubes
      0.3 ms  Reexport
    302.1 ms  Colors
    117.0 ms  ChainRulesCore
      0.7 ms  OpenLibm_jll
      0.3 ms  JLLWrappers
      6.7 ms  OpenSpecFun_jll 90.29% compilation time
     23.9 ms  SpecialFunctions
      0.3 ms  TensorCore
    253.4 ms  ColorVectorSpace 4.99% compilation time (100% recompilation)
     11.1 ms  ColorSchemes
      0.3 ms  NaNMath
      1.7 ms  Contour
   1781.9 ms  UnicodePlots
  4.167651 seconds (5.65 M allocations: 371.606 MiB, 6.49% gc time, 5.21% compilation time: 20% of which was recompilation)
```

**master**
```julia
julia> @time @time_imports using UnicodePlots
     34.1 ms  Preferences
      0.4 ms  SnoopPrecompile
      1.1 ms  Statistics
      4.5 ms  StaticArraysCore
   1060.5 ms  StaticArrays
     67.0 ms  FixedPointNumbers
    133.2 ms  ColorTypes 22.10% compilation time (100% recompilation)
     36.6 ms  Crayons
      1.8 ms  ConstructionBase
   1233.0 ms  Unitful
      1.9 ms  DataAPI
      0.2 ms  Compat
     11.1 ms  OrderedCollections
    111.3 ms  DataStructures
      0.6 ms  SortingAlgorithms
     15.0 ms  Missings
      1.6 ms  DocStringExtensions
      7.2 ms  IrrationalConstants
      0.4 ms  LogExpFunctions
      0.3 ms  StatsAPI
     36.0 ms  StatsBase
     86.5 ms  MarchingCubes
      0.4 ms  Reexport
    313.4 ms  Colors
    119.2 ms  ChainRulesCore
      0.5 ms  OpenLibm_jll
      0.4 ms  JLLWrappers
      6.2 ms  OpenSpecFun_jll 87.53% compilation time
     25.9 ms  SpecialFunctions
      0.4 ms  TensorCore
    267.2 ms  ColorVectorSpace 4.17% compilation time (100% recompilation)
     12.3 ms  ColorSchemes
      0.4 ms  Requires
      0.3 ms  NaNMath
      1.9 ms  Contour
    242.5 ms  FileIO
      1.0 ms  Bzip2_jll
      0.3 ms  Zlib_jll
      0.6 ms  FreeType2_jll
      6.6 ms  CEnum
      8.7 ms  FreeType
   2003.9 ms  UnicodePlots
  6.176219 seconds (7.55 M allocations: 486.762 MiB, 8.17% gc time, 3.75% compilation time: 18% of which was recompilation)
```
</details>